+++ .env
@@ -1,1 +1,1 @@
-SERVER_PORT=4550
+SERVER_PORT=6789
