+++ Dockerfile
@@ -3,7 +3,7 @@ ARG APP_INSIGHTS_AGENT_VERSION=3.4.12
 FROM hmctspublic.azurecr.io/base/java:17-distroless
 
 COPY lib/applicationinsights.json /opt/app/
-COPY build/libs/spring-boot-template.jar /opt/app/
+COPY build/libs/hmi-libra-dtu.jar /opt/app/
 
-EXPOSE 4550
-CMD [ "spring-boot-template.jar" ]
+EXPOSE 6789
+CMD [ "hmi-libra-dtu.jar" ]
