+++ Jenkinsfile_CNP
@@ -3,7 +3,7 @@
 @Library("Infrastructure")
 
 def type = "java"
-def product = "rpe"
-def component = "demo"
+def product = "hmi"
+def component = "libra-dtu"
 
 withPipeline(type, product, component) {}
