+++ charts/hmi-libra-dtu/Chart.yaml
@@ -1,11 +1,11 @@
 apiVersion: v2
 appVersion: "1.0"
-description: A Helm chart for spring-boot-template App
-name: rpe-spring-boot-template
-home: https://github.com/hmcts/spring-boot-template
+description: A Helm chart for hmi-libra-dtu App
+name: hmi-hmi-libra-dtu
+home: https://github.com/hmcts/hmi-libra-dtu
 version: 0.0.6
 maintainers:
-  - name: HMCTS rpe team
+  - name: HMCTS hmi team
 dependencies:
   - name: java
     version: 4.1.1
