+++ charts/hmi-libra-dtu/templates/NOTES.txt
