+++ charts/hmi-libra-dtu/values.yaml
@@ -1,0 +1,12 @@
+java:
+  applicationPort: 6789
+  image: 'hmctspublic.azurecr.io/hmi/libra-dtu:latest'
+  ingressHost: hmi-libra-dtu-{{ .Values.global.environment }}.service.core-compute-{{ .Values.global.environment }}.internal
+  aadIdentityName: hmi
+#  Uncomment once the vault containing the app insights key has been set up
+#  keyVaults:
+#    hmi:
+#      secrets:
+#        - name: AppInsightsInstrumentationKey
+#          alias: azure.application-insights.instrumentation-key
+  environment:
