+++ docker-compose.yml
@@ -1,7 +1,7 @@
 version: '2.1'
 
 services:
-  spring-boot-template:
+  hmi-libra-dtu:
     build:
       context: .
       args:
