+++ lib/applicationinsights.json
@@ -1,6 +1,6 @@
 {
-  "connectionString": "${file:/mnt/secrets/rpe/app-insights-connection-string}",
+  "connectionString": "${file:/mnt/secrets/hmi/app-insights-connection-string}",
   "role": {
-    "name": "rpe-demo"
+    "name": "hmi-libra-dtu"
   }
 }
