+++ src/functionalTest/java/uk/gov/hmcts/reform/hmi/controllers/SampleFunctionalTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.demo.controllers;
+package uk.gov.hmcts.reform.hmi.controllers;
 
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
@@ -12,7 +12,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import static io.restassured.RestAssured.given;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
-class SampleSmokeTest {
+class SampleFunctionalTest {
     protected static final String CONTENT_TYPE_VALUE = "application/json";
 
     @Value("${TEST_URL:http://localhost:8080}")
@@ -25,7 +25,7 @@ class SampleSmokeTest {
     }
 
     @Test
-    void smokeTest() {
+    void functionalTest() {
         Response response = given()
             .contentType(ContentType.JSON)
             .when()
@@ -35,5 +35,5 @@ class SampleSmokeTest {
 
         Assertions.assertEquals(200, response.statusCode());
         Assertions.assertTrue(response.asString().startsWith("Welcome"));
-    } 
+    }
 }
