+++ src/integrationTest/java/uk/gov/hmcts/reform/hmi/openapi/OpenAPIPublisherTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.demo.openapi;
+package uk.gov.hmcts.reform.hmi.openapi;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
