+++ src/main/resources/application.yaml
@@ -1,5 +1,5 @@
 server:
-  port: 4550
+  port: 6789
 
 management:
   endpoint:
@@ -12,13 +12,13 @@ management:
         include: health, info, prometheus
 
 springdoc:
-  packagesToScan: uk.gov.hmcts.reform.demo.controllers
+  packagesToScan: uk.gov.hmcts.reform.libra-dtu.controllers
   writer-with-order-by-keys: true
 
 #If you use a database then uncomment below lines and update db properties accordingly
 spring:
   config:
-    import: "optional:configtree:/mnt/secrets/rpe/"
+    import: "optional:configtree:/mnt/secrets/hmi/"
   application:
     name: Spring Boot Template
 #  datasource:
@@ -45,4 +45,4 @@ spring:
 
 azure:
   application-insights:
-    instrumentation-key: ${rpe.AppInsightsInstrumentationKey:00000000-0000-0000-0000-000000000000}
+    instrumentation-key: ${hmi.AppInsightsInstrumentationKey:00000000-0000-0000-0000-000000000000}
